+++ problem/problem_common.go
@@ -1,8 +1,6 @@
 package problem
 
 import (
-	"github.com/roidaradal/fn/ds"
-	"github.com/roidaradal/fn/list"
 	"github.com/roidaradal/opt/data"
 	"github.com/roidaradal/opt/discrete"
 	"github.com/roidaradal/opt/fn"
@@ -54,31 +52,5 @@ func newGraphCoverProblem(name string, variablesFn data.GraphVariablesFn) (*disc
 	}
 
 	p.Goal = discrete.Minimize
-	return p, graph
-}
-
-// Common steps for creating Independent Set problem
-func newIndependentSetProblem(name string) (*discrete.Problem, *data.Graph) {
-	p, graph := newGraphSubsetProblem(name, data.GraphVertices)
-	if p == nil || graph == nil {
-		return nil, nil
-	}
-
-	p.AddUniversalConstraint(func(solution *discrete.Solution) bool {
-		// Check that subset of vertices forms an independent set:
-		// none of the vertices are connected to each other
-		vertices := list.MapList(fn.AsSubset(solution), graph.Vertices)
-		// IsIndependentSet check
-		vertexSet := ds.SetFrom(vertices)
-		for _, vertex := range vertices {
-			adjacent := ds.SetFrom(graph.Neighbors(vertex))
-			if vertexSet.Intersection(adjacent).NotEmpty() {
-				return false
-			}
-		}
-		return true
-	})
-
-	p.Goal = discrete.Maximize
 	return p, graph
 }